refactor: move `cli/japa-plugin` to `src/japa-plugin`
